# Add CoveragePolygonValidator section performance optimization