Improves the performance of `CoveragePolygonValidator` and `CoverageValidator` by adding logic to avoid testing sections of ring vertices if the entire section does not overlap the target adjacent polygon.

## Performance comparison:

**World Polygons G99_5** 
Statistics: 323 geometries, 329,373 polygons, 459,540 vertices
Old code: 2646 ms
New code: 937 ms

**World Polygons G65**
Statistics: 323 geometries, 319,646 polygons, 27,706,942 vertices
Old code: 11809 s
New code: 103 s